refactor(interpreter): split REPLs out of lisp.ts
lisp.ts mixed the pure language (Interp, run, str, prelude, reader) with
two REPL front-ends and the process I/O plumbing. Keep lisp.ts
language-only and move the drivers into their own modules:

- repl.ts: reusable in-memory REPLs — MemoryRepl (eval/reset) and
  AgentRepl (adds halt + read-only conversation-state globals via
  setConversationVars, refreshed by the embedder each step).
- cli.ts: the interactive stdin/stdout loop and main() behind `pnpm repl`.

lisp.ts now exports Reader/EndOfFile (needed by cli.ts) and gains setExit
mirroring setWriter, so the CLI can wire process.exit without lisp.ts
importing node process I/O.
