+++ packages/interpreter/src/repl.ts
@@ -1,0 +1,190 @@
+/**
+ * Reusable REPL front-ends for the Lisptc interpreter.
+ *
+ * The interpreter (`./lisp.ts`) is pure language: `Interp`, `run`, `str`,
+ * `prelude`, the reader. A REPL is a driver on top of it. This module hosts the
+ * embeddable (string-in / string-out) REPLs; the interactive stdin/stdout loop
+ * lives in `./cli.ts`.
+ *
+ * - `MemoryRepl` — evaluate a program string, return what the interactive REPL
+ *   would have printed (last value + side-effect output, errors rendered
+ *   inline). No process I/O, so embedders can run Lisp without a subprocess.
+ * - `AgentRepl` — `MemoryRepl` plus the two things the pi coding-agent needs:
+ *   the `halt` built-in (a REPL feature, not part of the language) and
+ *   read-only conversation-state globals refreshed from the host each step.
+ */
+
+import {
+	Cell,
+	EndOfFile,
+	EvalException,
+	Interp,
+	type List,
+	newSym,
+	prelude,
+	run,
+	setWriter,
+	str,
+} from "./lisp.ts";
+
+// A REPL owns an interpreter and can be reset to a fresh one.
+export interface Repl {
+	readonly interp: Interp;
+	reset(): void;
+}
+
+// A REPL driven by strings rather than process I/O.
+export interface InMemoryRepl extends Repl {
+	eval(code: string): string;
+}
+
+// Convert a host JS value into the equivalent Lisp value. Mirrors `jsonToLisp`
+// in `./mcp.ts` (kept separate so the interpreter core stays free of it):
+// arrays -> proper list; plain objects -> alist of (key . value) with string
+// keys; null/undefined/false -> nil; true -> t; numbers/strings pass through.
+function jsToLisp(value: unknown): unknown {
+	if (value === null || value === undefined) return null;
+	if (value === true) return true;
+	if (value === false) return null; // Lisp has only nil for falsity
+	if (typeof value === "number" || typeof value === "bigint") return value;
+	if (typeof value === "string") return value;
+	if (Array.isArray(value)) return arrayToList(value.map(jsToLisp));
+	if (typeof value === "object") {
+		const pairs = Object.entries(value as Record<string, unknown>).map(
+			([k, v]) => new Cell(k, jsToLisp(v)),
+		);
+		return arrayToList(pairs);
+	}
+	return String(value);
+}
+
+function arrayToList(arr: unknown[]): List {
+	let out: List = null;
+	for (let i = arr.length - 1; i >= 0; i--) out = new Cell(arr[i], out);
+	return out;
+}
+
+// A persistent in-process REPL: `eval` runs a program and returns what the
+// interactive REPL would have printed. State persists across calls.
+export class MemoryRepl implements InMemoryRepl {
+	private currentInterp: Interp;
+
+	constructor() {
+		this.currentInterp = this.freshInterp();
+	}
+
+	get interp(): Interp {
+		return this.currentInterp;
+	}
+
+	private freshInterp(): Interp {
+		const interp = new Interp();
+		run(interp, prelude);
+		this.setup(interp);
+		return interp;
+	}
+
+	// Hook for subclasses to install extra bindings on a fresh interp. Called
+	// during construction and on every `reset()` — before subclass field
+	// initializers run on the first call, so overrides must tolerate their own
+	// fields still being undefined.
+	protected setup(_interp: Interp): void {}
+
+	// Evaluate a program; Lisp errors are rendered into the returned output
+	// rather than thrown.
+	eval(code: string): string {
+		let out = "";
+		const prev = setWriter((s) => {
+			out += s;
+		});
+		try {
+			// Bind value before appending: `out += str(run(...))` reads `out`
+			// before run() writes side-effect output into it, clobbering it.
+			const value = str(run(this.currentInterp, code));
+			out += `${value}\n`;
+		} catch (ex) {
+			if (ex instanceof EvalException) out += `${ex}\n`;
+			else if (ex === EndOfFile)
+				out += "unbalanced expression (unexpected end of input)\n";
+			else throw ex;
+		} finally {
+			setWriter(prev);
+		}
+		return out;
+	}
+
+	// Discard all definitions; start from a fresh prelude-loaded interp.
+	reset(): void {
+		this.currentInterp = this.freshInterp();
+	}
+}
+
+// The REPL the pi extension embeds: adds `halt` and the read-only conversation
+// globals (`conversation`, `user-messages`, `assistant-messages`).
+export class AgentRepl extends MemoryRepl {
+	// Set by the `halt` built-in this REPL installs; read (and cleared) via
+	// takeHalted(). `halt` is a REPL feature, not part of the language — a driver
+	// looping over eval() uses it to know when to stop.
+	private halted = false;
+	// The host-supplied conversation snapshot, kept so a post-error `reset()`
+	// can re-inject the globals until the next refresh.
+	private readonly conversationVars = new Map<string, unknown>();
+
+	// Re-establish the halt built-in and the conversation globals on every fresh
+	// interp. Guards `conversationVars` because the base constructor calls this
+	// before this subclass's field initializers have run.
+	protected override setup(interp: Interp): void {
+		this.installHalt(interp);
+		const vars = this.conversationVars;
+		if (vars) for (const [name, value] of vars) defineVar(interp, name, value);
+	}
+
+	// Install `(halt [value])`: record that the program asked to stop and return
+	// `value` (or t). It only sets a flag — evaluation of any following forms
+	// continues — so it is meant as the final form of a program.
+	private installHalt(interp: Interp): void {
+		const halt = interp.makeBuiltIn("halt", -1, (frame) => {
+			this.halted = true;
+			const rest = frame[0] as List;
+			return rest === null ? true : rest.car;
+		});
+		interp.defineGlobal(newSym("halt"), halt, {
+			signature: "(halt [value])",
+			doc: "Signal the REPL to stop after this program; return `value` (or t).",
+		});
+	}
+
+	// Replace the read-only conversation globals from a fresh host snapshot. Each
+	// entry (e.g. `conversation`, `user-messages`) becomes a global holding the
+	// Lisp translation of its JS value. Called before every eval, so the agent
+	// always sees the current transcript; a `(setq conversation …)` in user code
+	// survives only until the next refresh.
+	setConversationVars(vars: Record<string, unknown>): void {
+		this.conversationVars.clear();
+		for (const [name, value] of Object.entries(vars)) {
+			this.conversationVars.set(name, value);
+			defineVar(this.interp, name, value);
+		}
+	}
+
+	// Whether the last-evaluated program called `(halt)`; reads and clears the
+	// flag so each check reflects only evaluations since the previous check.
+	takeHalted(): boolean {
+		const h = this.halted;
+		this.halted = false;
+		return h;
+	}
+
+	override reset(): void {
+		super.reset();
+		this.halted = false;
+	}
+}
+
+// Define a single read-only conversation global on the given interp.
+function defineVar(interp: Interp, name: string, value: unknown): void {
+	interp.defineGlobal(newSym(name), jsToLisp(value), {
+		signature: name,
+		doc: "Read-only live conversation state (auto-updated each step).",
+	});
+}
